+++ desktop/applis/apps/bean/bean_core/src/test/java/code/formathtml/sample/SampleBeanOneMap.java
@@ -7,9 +7,9 @@ import code.util.StringMap;
 public class SampleBeanOneMap implements NatCaller {
     @Override
     public NaSt re(NaSt _instance, NaSt[] _args) {
-        StringMap<Integer> arr_ = new StringMap<Integer>();
-        arr_.addEntry("ONE",1);
-        arr_.addEntry("TWO",2);
+        StringMap<Long> arr_ = new StringMap<Long>();
+        arr_.addEntry("ONE",1L);
+        arr_.addEntry("TWO",2L);
         return BeanNatCommonLgNames.getStrInteger(arr_);
     }
 }
